considering a useless edge case for test directory deletation

+++ AbanLibTests/facility/Test.cs
@@ -17,6 +17,9 @@ public abstract class Test : IDisposable
 	public virtual void Dispose()
 	{
 		GC.SuppressFinalize(this);
-		Directory.Delete(DirectoryPath, true);
+		if (Directory.Exists(DirectoryPath))
+		{
+			Directory.Delete(DirectoryPath, true);
+		}
 	}
 }
